requirements.txt
langchain-astradb
langchain 
langchain-google-genai
langchain-openai
datasets
pypdf
python-dotenv
flask

-e .